correct errors

+++ makesymlinks.sh
@@ -6,8 +6,10 @@ cd $dir
 # if [[ "$OSTYPE"]] == "darwin"* ]]; then
 # fi
 
-echo "rm ~/.config/nvim"
-rm ~/.config/nvim
+if [ -d "~/.config/nvim"]; then
+    echo "rm -rf ~/.config/nvim"
+    rm -rf ~/.config/nvim
+fi
 
 echo "creat symlink to ~/.config/nvim"
 ln -s $dir/nvim ~/.config/nvim
@@ -18,8 +20,11 @@ echo ""
 files="bashrc vimrc"
 
 for file in $files; do
-    echo "rm $file from ~"
-    rm ~/.$file
+
+    if [-a "~/.$file"]; then
+        echo "rm $file from ~"
+        rm ~/.$file
+    fi
     echo "creat symlink to $file"
     ln -s $dir/$file ~/.$file
     echo ""
@@ -32,31 +37,47 @@ if [[ "$OSTYPE"]] == "linux-gnu"* ]]; then
     echo "-----"
     echo ""
 
-    mkdir -p ~/.local/bin
-    echo "create folders ~/.local/bin"
+    if [ ! -d "~/.local/bin"]; then
+        mkdir -p ~/.local/bin
+        echo "create folders ~/.local/bin"
+    fi
 
-    mkdir -p ~/.local/share/themes
-    echo "create folders ~/.local/share/themes"
+    if [ ! -d "~/.local/share/themes"]; then
+        mkdir -p ~/.local/share/themes
+        echo "create folders ~/.local/share/themes"
+    fi
 
-    mkdir -p ~/.local/share/icons
-    echo "create folders ~/.local/share/icons"
+    if [ ! -d "~/.local/share/icons"]; then
+        mkdir -p ~/.local/share/icons
+        echo "create folders ~/.local/share/icons"
+    fi
 
-    mkdir -p ~/.local/share/fonts
-    echo "create folders ~/.local/share/fonts"
+    if [ ! -d "~/.local/share/fonts"]; then
+        mkdir -p ~/.local/share/fonts
+        echo "create folders ~/.local/share/fonts"
+    fi
 
-    rm ~/.local/bin/nvim
+    if [ -f "~/.local/bin/nvim"]; then
+        rm ~/.local/bin/nvim
+    fi
     ln -s ~/Documents/nvim-linux64/bin/nvim ~/.local/bin/nvim
     echo "creat symlink to ~/.local/bin/nvim"
 
-    rm ~/.local/bin/tree-sitter
+    if [ -f "~/.local/bin/tree-sitter"]; then
+        rm ~/.local/bin/tree-sitter
+    fi
     ln -s $dir/bins_lin/tree-sitter ~/.local/bin/tree-sitter
-    chmod +x ~/.local/bin/tree-sitter
     echo "creat symlink to ~/.local/bin/tree-sitter"
 
-    rm ~/.local/bin/stylua
+    chmod +x ~/.local/bin/tree-sitter
+
+    if [ -f "~/.local/bin/stylua"]; then
+        rm ~/.local/bin/stylua
+    fi
     ln -s $dir/bins_lin/stylua ~/.local/bin/stylua
-    chmod +x ~/.local/bin/stylua
     echo "creat symlink to ~/.local/bin/stylua"
+
+    chmod +x ~/.local/bin/stylua
 
     # rm ~/.local/bin/ctags
 fi
